+++ group7/src/main/java/com/cpt202/group7/controller/HomeController.java
@@ -1,0 +1,12 @@
+package com.cpt202.group7.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+public class HomeController {
+    @RequestMapping("/home")
+    public String showHomePage(){
+        return "home";
+    }
+}
