+++ group7/src/main/resources/templates/api/auth/login.html
@@ -2,9 +2,12 @@
 <html>
 <head>
     <title>Login</title>
+    <link rel="shortcut icon" th:href="@{/imgs/icon/favicon.jpg}"/>
 </head>
 <body>
+<h1>This is Login Page</h1>
 <h2>Login Form</h2>
+
 <form action="/api/auth/login" method="post">
     <label for="username">Username:</label>
     <input type="text" id="username" name="username" required><br><br>
@@ -14,5 +17,10 @@
 
     <input type="submit" value="Login">
 </form>
+
+<form action="/home" method="get">
+    <input type="submit" value="Cancel">
+</form>
+
 </body>
 </html>
