--- group7/src/main/resources/templates/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>首页</title>
-</head>
-<body>
-<h1>欢迎来到首页</h1>
-</body>
-</html>
